fix validations
