+++ models/contacts.js
@@ -1,8 +1,19 @@
 const fs = require('fs/promises');
 const path = require('path');
 const { v1: uuidv1 } = require('uuid')
-
+const Joi = require('joi')
 const contactsPath = path.join(__dirname, 'contacts.json');
+const createContactShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2 }).required(),
+  phone: Joi.string().regex(/^[0-9]{10}$/).messages({ 'string.pattern.base': `Phone number must have 10 digits.` }).required()
+})
+const updateContactSchema = Joi.object({
+  name: Joi.string(),
+  email: Joi.string().email({ minDomainSegments: 2 }).allow('', null),
+  phone: Joi.string().allow('', null).regex(/^[0-9]{10}$/).messages({ 'string.pattern.base': `Phone number must have 10 digits.` })
+})
 const contactsListNormalization = async (contacts) => await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2), 'utf-8');
 
 
@@ -27,17 +38,27 @@ const removeContact = async (contactId) => {
 }
 
 const addContact = async (body) => {
-  const contacts = await listContacts();
-  const newContact = { id: uuidv1(), ...body };
-  contacts.push(newContact);
+  const contacts = await listContacts()
+  const { error } = createContactShema.validate(body, { abortEarly: false })
+  if (error) {
+    const errorMessage = error.details.map(detail => detail.message).join('; ')
+    throw new Error(errorMessage)
+  }
+  const newContact = { id: uuidv1(), ...body }
+  contacts.push(newContact)
   contactsListNormalization(contacts)
-  return newContact;
+  return newContact
 }
 
 const updateContact = async (contactId, body) => {
   const contacts = await listContacts()
   const index = contacts.findIndex((item) => item.id === contactId)
   if (index === -1) return null
+  const { error } = updateContactSchema.validate(body, { abortEarly: false })
+  if (error) {
+    const errorMessage = error.details.map(detail => detail.message).join('; ')
+    throw new Error(errorMessage)
+  }
   contacts[index] = { ...contacts[index], ...body }
   contactsListNormalization(contacts)
   return contacts[index]
