+++ models/contacts.json
@@ -70,5 +70,11 @@
     "name": "Nick 123456781345",
     "email": "exex@ex.ua",
     "phone": "12345612345"
+  },
+  {
+    "id": "e4233cb0-e5d6-11ed-a4ec-0b332dc51b45",
+    "name": "Bill upd",
+    "email": "exexexe@ex.com",
+    "phone": "1234567891"
   }
 ]
